+++ server-adapters/hono/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mastra/hono
 
+## 1.1.2-alpha.0
+
+### Patch Changes
+
+- Updated dependencies [[`90f7894`](https://github.com/mastra-ai/mastra/commit/90f7894568dc9481f40a4d29672234fae23090bb), [`8109aee`](https://github.com/mastra-ai/mastra/commit/8109aeeab758e16cd4255a6c36f044b70eefc6a6), [`8109aee`](https://github.com/mastra-ai/mastra/commit/8109aeeab758e16cd4255a6c36f044b70eefc6a6), [`d917195`](https://github.com/mastra-ai/mastra/commit/d917195995422dff39ee46a516fe7f7205158858)]:
+  - @mastra/core@1.2.1-alpha.0
+  - @mastra/server@1.2.1-alpha.0
+
 ## 1.1.1
 
 ### Patch Changes